# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pokemon": {
      "post": {
        "operationId": "PokemonController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePokemonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pokemon created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin role required"
          },
          "409": {
            "description": "Pokemon already exists"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new pokemon (Admin only)",
        "tags": [
          "pokemon"
        ]
      },
      "get": {
        "operationId": "PokemonController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pokemon retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PokemonResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all pokemon (no pagination, cached)",
        "tags": [
          "pokemon"
        ]
      }
    },
    "/pokemon/search": {
      "get": {
        "operationId": "PokemonController_search",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "enum": [
                "name",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter Pokemon by name (partial match, case-insensitive)",
            "schema": {
              "example": "pikachu",
              "type": "string"
            }
          },
          {
            "name": "types",
            "required": false,
            "in": "query",
            "description": "Comma-separated Pokemon types",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "Bug",
                  "Dark",
                  "Dragon",
                  "Electric",
                  "Fairy",
                  "Fighting",
                  "Fire",
                  "Flying",
                  "Ghost",
                  "Grass",
                  "Ground",
                  "Ice",
                  "Normal",
                  "Poison",
                  "Psychic",
                  "Rock",
                  "Steel",
                  "Water"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pokemon retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPokemonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Search Pokemon with pagination",
        "tags": [
          "pokemon"
        ]
      }
    },
    "/pokemon/count": {
      "get": {
        "operationId": "PokemonController_getCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pokemon count retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonCountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get total Pokemon count in the system",
        "tags": [
          "pokemon"
        ]
      }
    },
    "/pokemon/{id}": {
      "get": {
        "operationId": "PokemonController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pokemon ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pokemon retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pokemon not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get pokemon by ID",
        "tags": [
          "pokemon"
        ]
      },
      "patch": {
        "operationId": "PokemonController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pokemon ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePokemonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pokemon updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin role required"
          },
          "404": {
            "description": "Pokemon not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update pokemon (Admin only)",
        "tags": [
          "pokemon"
        ]
      },
      "delete": {
        "operationId": "PokemonController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pokemon ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pokemon deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin role required"
          },
          "404": {
            "description": "Pokemon not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete pokemon (Admin only)",
        "tags": [
          "pokemon"
        ]
      }
    },
...
    "/upload/image": {
      "post": {
        "operationId": "UploadController_uploadImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (JPG, PNG, GIF, WebP, max 5MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or upload failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin role required"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Upload an image to Cloudinary (Admin only)",
        "tags": [
          "upload"
        ]
      }
    }
  },
  "info": {
    "title": "Pokemon Ranking API",
    "description": "API for managing Pokemon and user rankings",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "PokemonResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "example": "Pikachu"
          },
          "image": {
            "type": "string",
            "example": "https://example.com/pokemon/pikachu.png"
          },
          "types": {
            "type": "array",
            "example": [
              "Electric"
            ],
            "items": {
              "type": "string",
              "enum": [
                "Bug",
                "Dark",
                "Dragon",
                "Electric",
                "Fairy",
                "Fighting",
                "Fire",
                "Flying",
                "Ghost",
                "Grass",
                "Ground",
                "Ice",
                "Normal",
                "Poison",
                "Psychic",
                "Rock",
                "Steel",
                "Water"
              ]
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-20T15:45:00.000Z"
          }
        },
        "required": [
          "_id",
          "name",
          "image",
          "types",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreatePokemonDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Pikachu"
          },
          "image": {
            "type": "string",
            "example": "pikachu.png",
            "description": "Pokemon image - either a filename (e.g., \"pikachu.png\") or a URL from whitelisted domains"
          },
          "types": {
            "type": "array",
            "example": [
              "Electric"
            ],
            "description": "Pokemon types",
            "items": {
              "type": "string",
              "enum": [
                "Bug",
                "Dark",
                "Dragon",
                "Electric",
                "Fairy",
                "Fighting",
                "Fire",
                "Flying",
                "Ghost",
                "Grass",
                "Ground",
                "Ice",
                "Normal",
                "Poison",
                "Psychic",
                "Rock",
                "Steel",
                "Water"
              ]
            }
          }
        },
        "required": [
          "name",
          "image"
        ]
      },
      "PaginatedPokemonResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PokemonResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 100,
            "description": "Total number of Pokemon"
          },
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "number",
            "example": 20,
            "description": "Number of items per page"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "PokemonCountResponseDto": {
        "type": "object",
        "properties": {
          "totalPokemonCount": {
            "type": "number",
            "example": 1025,
            "description": "Total Pokemon registered in the system"
          }
        },
        "required": [
          "totalPokemonCount"
        ]
      },
      "UpdatePokemonDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Pikachu"
          },
          "image": {
            "type": "string",
            "example": "pikachu.png",
            "description": "Pokemon image - either a filename (e.g., \"pikachu.png\") or a URL from whitelisted domains"
          },
          "types": {
            "type": "array",
            "example": [
              "Electric"
            ],
            "description": "Pokemon types",
            "items": {
              "type": "string",
              "enum": [
                "Bug",
                "Dark",
                "Dragon",
                "Electric",
                "Fairy",
                "Fighting",
                "Fire",
                "Flying",
                "Ghost",
                "Grass",
                "Ground",
                "Ice",
                "Normal",
                "Poison",
                "Psychic",
                "Rock",
                "Steel",
                "Water"
              ]
            }
          }
        }
      },
...
      "UploadResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "The URL of the uploaded image",
            "example": "https://res.cloudinary.com/your-cloud/image/upload/v1234567890/pokemon/abc123.png"
          }
        },
        "required": [
          "url"
        ]
      }
    }
  }
}